requirements.txt
asgiref==3.8.1
black==24.4.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
filelock==3.15.4
gunicorn==22.0.0
identify==2.5.36
iniconfig==2.0.0
isort==5.13.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pytest==8.2.2
pytest-django==4.8.0
PyYAML==6.0.1
sqlparse==0.5.0
virtualenv==20.26.3
